fix: CLOCK_REALTIME.toUInt() for Kotlin/Native POSIX clock_gettime binding

+++ persistence/src/iosMain/kotlin/io/neuralheads/kmpworker/persistence/CurrentEpochMillis.kt
@@ -15,6 +15,6 @@ import platform.posix.timespec
  */
 actual fun currentEpochMillis(): Long = memScoped {
     val ts = alloc<timespec>()
-    clock_gettime(CLOCK_REALTIME, ts.ptr)
+    clock_gettime(CLOCK_REALTIME.toUInt(), ts.ptr)
     ts.tv_sec * 1_000L + ts.tv_nsec / 1_000_000L
 }
